Bug fix with len

+++ template_one_case.cpp
@@ -24,7 +24,7 @@ using namespace std;
 #define pqll priority_queue<ll>
 
 #define all(x) (x).begin(), (x).end()
-#define len(x) (ll((x).size()))
+#define len(x) (int((x).size())) // need int here, otherwise bugs
 #define forn(i,n) for(ll i = 0; i < n ; i++)
 
 #define debug(x) cout << #x << " = " << x << "\n";
